+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.0.3 (Released 2015/09/28)
+* Added current instance aliase
+* Remove duplicate settings
+
 ### New in 0.0.2 (Released 2015/09/23)
 * New extensions methods